# Gateway/openapi.yaml
openapi: 3.0.0
info:
  title: Gateway API
  description: Point d'entrée principal pour l'architecture microservices du jeu Micro-Donjon. Orchestre les appels vers tous les services.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Serveur local de la Gateway
paths:
...
  /api/logs:
    post:
      summary: Créer un nouvel événement de log
      tags:
        - Logs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogCreate'
      responses:
        '201':
          description: Log créé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Log'
        '400':
          description: Données invalides
        '500':
          description: Erreur interne du serveur

    get:
      summary: Récupérer les logs avec pagination et filtres
      tags:
        - Logs
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: level
          in: query
          schema:
            type: string
            enum: [DEBUG, INFO, WARN, ERROR]
        - name: service
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Liste paginée des logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Log'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer
        '500':
          description: Erreur interne du serveur
...
components:
  schemas:
...
    Log:
      type: object
      required:
        - id
        - level
        - message
        - service
        - timestamp
      properties:
        id:
          type: string
          format: uuid
        level:
          type: string
          enum: [DEBUG, INFO, WARN, ERROR]
        message:
          type: string
        service:
          type: string
        userId:
          type: integer
        sessionId:
          type: string
        metadata:
          type: object
        timestamp:
          type: string
          format: date-time
    LogCreate:
      type: object
      required:
        - level
        - message
        - service
      properties:
        level:
          type: string
          enum: [DEBUG, INFO, WARN, ERROR]
        message:
          type: string
        service:
          type: string
        userId:
          type: integer
        sessionId:
          type: string
        metadata:
          type: object
...